Move categories editor to a saparate package
