+++ app/src/main/java/com/yurii/youtubemusic/utilities/Injector.kt
@@ -7,6 +7,7 @@ import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.yurii.youtubemusic.services.mediaservice.MediaService
 import com.yurii.youtubemusic.services.mediaservice.MusicServiceConnection
 import com.yurii.youtubemusic.models.Category
+import com.yurii.youtubemusic.screens.categories.CategoriesEditorViewModel
 import com.yurii.youtubemusic.screens.saved.SavedMusicViewModel
 import com.yurii.youtubemusic.screens.youtube.YouTubeMusicViewModel
 import com.yurii.youtubemusic.viewmodels.*
@@ -38,6 +39,11 @@ object Injector {
     fun provideYouTubeMusicViewModel(context: Context, googleSignInAccount: GoogleSignInAccount): YouTubeMusicViewModel.Factory {
         val preferences = ServiceLocator.providePreferences(context)
         return YouTubeMusicViewModel.Factory(context, googleSignInAccount, preferences)
+    }
+
+    fun provideCategoriesEditorViewMode(context: Context): CategoriesEditorViewModel.Factory {
+        val preferences = ServiceLocator.providePreferences(context)
+        return CategoriesEditorViewModel.Factory(preferences)
     }
 
     private fun provideMusicServiceConnection(context: Context): MusicServiceConnection {
