+++ app/src/main/res/layout/activity_categories_editor.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".CategoriesEditorActivity">
+    tools:context=".screens.categories.CategoriesEditorActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
